+++ frontend/VibeVault/src/App.css
@@ -14,6 +14,28 @@ h1{
     
 }
 
+#Upload{
+    font-family: monospace;
+    color: rgb(0, 255, 26);
+    font-size: 200px;
+    display: flex;
+    justify-content: center;
+    text-align: center;
+    align-items: center;
+    font-size: 50px;
+    
+}
+
+#Uploadcircle{
+    width: 300px; 
+    height: 100px;
+    border-radius: 50%; 
+    background-color: rgba(255, 255, 255, 0.3);
+    position: absolute;
+    left: 585px;
+    bottom: 70px;
+}
+
 .navbar {
     overflow: hidden;
     height: 7q0px;
