+++ frontend/VibeVault/src/Components/Upload.jsx
@@ -1,6 +1,4 @@
 import { useState } from 'react'
-import "../Components/secondpage.css"
-import image1 from "../../assets/Group 7.png"
 
 
 
@@ -9,8 +7,8 @@ function App() {
   
     return (
       <>
-        <img id="topic" src={image1} alt="Example" />
-        
+      <div id="Uploadcircle"></div>
+      <div id='Upload'>UPLOAD</div> 
 
       </>
     )
